feat(projects): add the sort cabins by name, price, and quantity feature
Implement the sort feature on the cabins page. Cabins can be sorted by
name, price, and quantity in ascending and descending order.

Convert the `useFilterCabins` to `useQueryOperation` hook to manage all
kind of query operation, such as filtering or sorting.

Add an `arg` property to the `Filter.Option` to allow conditional
filtering.