feat(projects): add a new route to see details of a booking
Add a new route to the app router to see details of a booking. Create a
new page component called `Booking` to display the booking details. The
`Booking` has a `BookingDetails` component that showing necessary
booking information. The `BookingDetails` has a presentational component
called `BookingDataBox` which shows all focal data of a booking. Add an
action column to the `BookingRow` component to take necessary actions as
needed. Create a `getBooking` api function to fetch data of a booking by
its id from the supabse.