feat(projects): add a new page to check in booking
Create a `Checkin` page component and configure it in the app router.
The page has a `CheckinBooking` component that displays all booking
information. This component has a `Checkbox` for guests payment
confirmation. Without payment confirmation, users can't check in guests
booking. Add a menu button to the `BookingRow` action column to navigate
the checkin page.

* add a `Checkin` page component
* add a `CheckinBooking` component
* add a `useCheckin` hook to mutate booking for check-in
* add a `updateBooking` api function to update a booking by its id