feat(projects): allow users to add breakfast at booking check-in time
Add a `Checkbox` to the `CheckinBooking` component to allow users to
add breakfast by checking the checkbox. Revise the `useCheckin` hook to
update the booking total price, breakfast price, and breakfast status.
If users check the checkbox, then the booking total price and
breakfast price are recalculate and it's updated when check-in process
is complete.