feat(projects): implement the delete a booking feature
The feature to delete a booking has been implemented on the bookings page.
A `useDeleteBooking` hook is created to delete a booking. This hook takes
booking id as a parameter and returns an object containing methods and
properties. The return has a `bookingDeleteTrigger` method to trigger
the booking delete mutation. Add a `ConfirmDeleteModal` on the `BookingRow`
action column to take the delete booking action with an alert message.

* Add a `useDeleteBooking` hook function to manage the delete booking mutation