feat(projects): implement the user login feature with login page
The login page is implemented with the `LoginForm` component. The
`LoginForm` contains all the logic of the user login feature. Create a
`useLogin` hook function with `loginFetcher` api function to manage the
user login. `useLogin` returns an object containing `loginTrigger` method.
This method takes user credentials as a parameter to complete the user
login process. Create a `LoginSchema` to validate the login form.