feat(projects): implement user logout functionality
Create a `LogoutButton` component and add it to the app header to initiate
the logout process. Create a `useLogout` hook with the `logoutFetcher` api
function to make the logout process manageable. The `logoutFetcher` logs
out the logged-in user and clears the session. The `useLogout` hook
returns an object containing a `handleLogout` method and this method calls
the `logoutFetcher` function. The `useLogout` hook manage the logout
mutation with loading and error states. The user is redirected to the login
page upon successful logout.