feat(projects): implement create a new user functionality
Create a `SignUpForm` with name, email, password, and confirm-password
fields to register a new user. Add a `useSignUp` hook to manage the
signup mutation. This hook returns an object contains a `signupTrigger`
method. This method actually triggers the `signupFetcher` api function
with new user credentials to register a user to the server. Clear the
form fields and notify with a success toast upon successful singup.