feat(projects): add a `Previewer` component to preview the file input image
The `Previewer` component previews an image from a file input field. It
takes name of the file input field as a prop and uses the `useFormikContext`
hook to retrieve the value of that input field. It contains a `Photo`
component. When users change the input value it reads that file and
converts it to a `DataURL` string. After that, it updates the `Photo`
component `src` prop with the `DataURL` to preview user's uploaded image.