feat(projects): implement the user data update functionality
Built the account page with a `UpdateUserDataForm` component. This
component is made to update user data. It is built upon Formik library.
This component contains a `Previewer` component to see the user's avatar.
The submit button of `UpdateUserDataForm` is only enabled when the form
filed values are valid and changed. Create a `UpdateUserSchema` to
validate this form and create a `updateLoggedInUser` api function to
update the logged-in user data. To manage user update mutation create a
`useUserMutation` hook. The `updateLoggedInUser` function updates either
the password nor the user's metadata at once.