feat(projects): implement the dark theme with css variables
Assign the default css variables to the app root element for the light
theme then reassign the css variables to a new selector using the root
element combined with a class named `dark` for the dark theme. Therefore,
to apply the dark theme to add the dark class into the app root element.

To manage the theme state create a `ThemeContext` and `ThemeProvider`.
The `ThemeProvider` provides an object containing theme property,
changeTheme, and handleToggleTheme methods. Create a `ThemeToggleButton`
to toggle the application theme between light and dark. This button is
added to the application header.