feat(projects): create a `useBookingStatistics` hook
The `useBookingStatistics` is made to get booking statistics to display
on the dashboard page. This hook is created with `getBookingStatistics`
fetcher function. This function brings all necessary data from the server
and computes booking statistics from the given date to the present day.
The function computes the sales amount, bookings count, confirm bookings
count, check-in bookings count, occupancy rate, and today's activities.