feat(projects/18): divide main layout between base and account layouts
This layout division is made to create a different layout for the account
page. The base layout contains homepage, about, and cabins pages. The
account layout contains the account and some other pages associated with
it. The account layout has a header that is fixed position at the top.
But the base layout header isn't fixed position at the top.